Add Starter Project: Laravel + Vue 3 Boilerplate
An implementation of the **Laravel Breeze** application / authentication starter kit frontend in Vue.js v3. How a library can become a framework with the help of **pinia** and **vue-router**. This repository emulates the same purpose achieved by Taylor Otwell with his demo repository "breeze-next", but with VueJs v3 instead of React.